requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-dialogflow-cx==1.35.0
googleapis-common-protos==1.65.0
idna==3.8
numpy==2.1.1
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
python-dotenv==1.0.1
rsa==4.9
setuptools==75.1.0
wheel==0.44.0
pyyaml
rospkg
